feat: 코디네이터 역할 분리 #42
