+++ Yut/Yut/APP/ARViewContainer.swift
@@ -14,8 +14,8 @@ struct ARViewContainer: UIViewRepresentable {
     @ObservedObject var arState: ARState // ContentView로부터 arState를 받아서 관찰
     
     // MARK: - Coordinator 생성
-    func makeCoordinator() -> Coordinator {
-        Coordinator()
+    func makeCoordinator() -> ARCoordinator {
+        ARCoordinator()
     }
     
     // swiftUI 뷰가 생성될 때 한번만 호출됨 -> ARView 초기화 & 설정
@@ -28,10 +28,11 @@ struct ARViewContainer: UIViewRepresentable {
         arView.session.delegate = context.coordinator       // AR 이벤트 수신
         
         // MARK: - Gesture Recognizers 설정
-        let tapGesture = UITapGestureRecognizer(target: context.coordinator, action: #selector(Coordinator.handleTap))
-        let pinchGesture = UIPinchGestureRecognizer(target: context.coordinator, action: #selector(Coordinator.handlePinch))
-        let panGesture = UIPanGestureRecognizer(target: context.coordinator, action: #selector(Coordinator.handlePan))
-        let rotationGesture = UIRotationGestureRecognizer(target: context.coordinator, action: #selector(Coordinator.handleRotation))
+        let gestureHandler = context.coordinator.gestureHandler!
+        let tapGesture = UITapGestureRecognizer(target: gestureHandler, action: #selector(GestureHandler.handleTap))
+        let pinchGesture = UIPinchGestureRecognizer(target: gestureHandler, action: #selector(GestureHandler.handlePinch))
+        let panGesture = UIPanGestureRecognizer(target: gestureHandler, action: #selector(GestureHandler.handlePan))
+        let rotationGesture = UIRotationGestureRecognizer(target: gestureHandler, action: #selector(GestureHandler.handleRotation))
         
         arView.addGestureRecognizer(tapGesture)
         arView.addGestureRecognizer(pinchGesture)
