--- Yut/Yut/APP/Coordinator/AppCoordinator.swift
@@ -1,1 +1,0 @@
-// 화면 흐름 관리
