+++ Yut/Yut/Features/GameLogicManager.swift
@@ -1,0 +1,13 @@
+//
+//  GameLogicManager.swift
+//  Yut
+//
+//  Created by yunsly on 7/21/25.
+//
+
+import Foundation
+
+class GameLogicManager {
+    // 게임 로직(턴 관리, 점수 계산 등)
+    
+}
